Fixes and improvements

+++ src/main/java/com/wojnarowicz/sfg/gw/api/model/sap/BCCoverageDTO.java
@@ -1,5 +1,7 @@
 package com.wojnarowicz.sfg.gw.api.model.sap;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -19,7 +21,7 @@ import lombok.Setter;
 public class BCCoverageDTO {
 
     @JsonProperty("ChargeAmount")
-    public Float chargeAmount;
+    public BigDecimal chargeAmount;
     
     @JsonProperty("ChargeCode")
     public String chargeCode;
